Comment to story 3
It does not made sense that you want to enter the forest and you have already meet goatkeeper (which means you are already IN the forest).
if creating dependent stories please consider all the possibilities from the previous one to make consistent story. If in the Story 1 -Goats user selected option 1 and killed one of the goats, the goat-keeper does never appear. 

I also recommend some spell-check for story 3

Also logic in the robbery part has problem: if you fight off the attacker but did not had a knife, does he also drops it then?